refactor: fold render and htmx into web (drop ui domain)
Follows the updated docs/domain-structure.md, which reclassifies render and
htmx as response-side web transport (symmetric to web/request) rather than a
standalone ui domain:

- ui/render -> web/render
- ui/htmx   -> web/htmx
- ui/ removed (no shipped packages remain)

Import paths rewritten (forge/ui/render -> forge/web/render, forge/ui/htmx ->
forge/web/htmx); package names and call sites unchanged. Shipped packages now
occupy six domains (core crypto resilience web data ops); status note updated.

Verified: go build, go vet, just lint (0 issues), just test -race (59 pkgs pass).
