+++ web/htmx/response_test.go
@@ -10,7 +10,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/dmitrymomot/forge/ui/htmx"
+	"github.com/dmitrymomot/forge/web/htmx"
 )
 
 func TestSimpleSetters(t *testing.T) {
