+++ web/render/stream.go
