+++ web/render/templ_test.go
@@ -11,7 +11,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/dmitrymomot/forge/ui/render"
+	"github.com/dmitrymomot/forge/web/render"
 )
 
 type testCtxKey struct{}
